+++ src/App.test.js
@@ -212,7 +212,7 @@ describe('Beta banner', () => {
     expect(screen.queryByText('betaBanner.message')).toBeInTheDocument()
   })
 
-  test('Does not render beta banner if betaBannerDismissedCookie is true', () => {
+  test('Does not render beta banner if betaBannerDismissed cookie is true', () => {
     cookies.set('betaBannerDismissed', 'true')
     render(
       <CookiesProvider cookies={cookies}>