+++ src/components/ProjectListItem/ProjectListItem.test.js
@@ -1,6 +1,7 @@
 import { fireEvent, render, screen } from "@testing-library/react";
 import React from "react";
 import { Provider } from "react-redux";
+import { MemoryRouter } from "react-router-dom";
 import configureStore from 'redux-mock-store';
 import ProjectListItem from "./ProjectListItem";
 
@@ -15,7 +16,7 @@ beforeEach(() => {
   const initialState = {}
   store = mockStore(initialState);
 
-  render(<Provider store={store}><ProjectListItem project = {project}/></Provider>)
+  render(<Provider store={store}><MemoryRouter><ProjectListItem project = {project}/></MemoryRouter></Provider>)
 })
 
 test('Renders project name', () => {