+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Simplified the URL structure to replace project type (#347)
 
+### Fixed
+
+- Show `p5` error messages in the user interface (#346)
+
 ## [0.11.0] - 2023-01-17
 
 ### Added