+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 
 - Show `p5` error messages in the user interface (#346)
+- Get review apps working (#351)
 
 ## [0.11.0] - 2023-01-17